opto stubs write thread anchor pc rather than expect ret pc to be on stack
the x86 opto stubs leave their VM callouts ot find their return
address on the stack above the therad anchor sp. this doesn't work on
AArch64 as the C compiler can put the ret pc where it likes. so we
ahve to write the anchor pc fiels form the opto stub. it's baroque but
it works.